+++ src/Bugsnag/Client.cs
@@ -162,9 +162,6 @@ namespace Bugsnag
 
       if (!report.Ignored)
       {
-        Bugsnag.InternalMiddleware.DetermineDefaultContext(report);
-        Bugsnag.InternalMiddleware.ApplyMetadataFilters(report);
-
         _delivery.Send(report);
 
         Breadcrumbs.Leave(Breadcrumb.FromReport(report));