+++ src/Bugsnag/Payload/App.cs
@@ -5,7 +5,7 @@ namespace Bugsnag.Payload
   /// <summary>
   /// Represents the "app" key in the error report payload.
   /// </summary>
-  public class App : Dictionary<string, object>
+  public class App : Dictionary<string, object>, IFilterable
   {
     public App(IConfiguration configuration) : this(configuration.AppVersion, configuration.ReleaseStage, configuration.AppType)
     {