+++ src/Bugsnag/Serializer.cs
@@ -1,4 +1,6 @@
+using System.Collections.Generic;
 using System.Diagnostics;
+using System.Linq;
 using System.Text;
 
 namespace Bugsnag
@@ -7,9 +9,18 @@ namespace Bugsnag
   {
     public static string SerializeObject(object obj)
     {
+      return SerializeObject(obj, new string[0]);
+    }
+
+    public static string SerializeObject(object obj, string[] filters)
+    {
+      var parsedFilters = filters != null ?
+        filters.ToDictionary(m => m, m => true) :
+        new Dictionary<string, bool>();
+
       try
       {
-        return SimpleJson.SimpleJson.SerializeObject(obj);
+        return SimpleJson.SimpleJson.SerializeObject(obj, parsedFilters);
       }
       catch (System.Exception exception)
       {
@@ -21,9 +32,14 @@ namespace Bugsnag
 
     public static byte[] SerializeObjectToByteArray(object obj)
     {
+      return SerializeObjectToByteArray(obj, new string[0]);
+    }
+
+    public static byte[] SerializeObjectToByteArray(object obj, string[] filters)
+    {
       try
       {
-        var serializedObject = SerializeObject(obj);
+        var serializedObject = SerializeObject(obj, filters);
         return Encoding.UTF8.GetBytes(serializedObject);
       }
       catch (System.Exception exception)