+++ CHANGELOG.md
@@ -1,6 +1,28 @@
 Changelog
 =========
 
+## Unreleased
+
+### Enhancements
+
+* Improve metadata filtering
+  | [martin308](https://github.com/martin308)
+  | [#90](https://github.com/bugsnag/bugsnag-dotnet/pull/90)
+
+### Bug fixes
+
+* Only send session data when it exists
+  | [martin308](https://github.com/martin308)
+  | [98](https://github.com/bugsnag/bugsnag-dotnet/pull/98)
+
+* Request context is now available in custom middleware
+  | [martin308](https://github.com/martin308)
+  | [#91](https://github.com/bugsnag/bugsnag-dotnet/pull/91)
+
+* Only set the context to the current request URL if the context has not already been set
+  | [tremlab](https://github.com/tremlab)
+  | [#89](https://github.com/bugsnag/bugsnag-dotnet/pull/89)
+
 ## 2.0.2 (2018-03-28)
 
 ### Bug fixes